#  Fix lazy state cloning so that it is actually lazy and optimize CardinalityConstraint