Door in the Archway (quadruple/LK/killer puzzle) timing actually got worse after fixing lazy state cloning (41s -> 44s), what had a bigger impact is the second commit to have an early exit in `enforce/enforceCandidateElim` (44s -> 30s).

I'm guessing that this is due to worse branch prediction, but not sure. Could be worth creating an "eagerly cloned constraint state", but I think it's probably right to keep the lazy cloning as a feature as constraints with very large states may benefit more from lazy cloning (e.g. perhaps OrConstraint)